Move plustek mock server to its own file
