+++ services/scan/src/index.ts
@@ -9,8 +9,9 @@ import {
   VX_MACHINE_TYPE,
 } from './globals';
 import { LoopScanner, parseBatchesFromEnv } from './loop_scanner';
-import { plustekMockServer, Scanner } from './scanners';
+import { Scanner } from './scanners';
 import * as server from './server';
+import { plustekMockServer } from './plustek_mock_server';
 
 const logger = new Logger(LogSource.VxScanService);
