+++ services/scan/src/scanners/plustek.ts
@@ -1,16 +1,13 @@
 import {
   ClientDisconnectedError,
-  MockScannerClient,
   PaperStatus,
   ScannerClient,
   ScannerError,
   ScanRetryPredicate,
 } from '@votingworks/plustek-sdk';
-import { ok, Provider, Result, safeParse } from '@votingworks/types';
+import { ok, Provider, Result } from '@votingworks/types';
 import { Scan } from '@votingworks/api';
 import makeDebug from 'debug';
-import express, { Application } from 'express';
-import * as z from 'zod';
 import { BatchControl, Scanner, ScanOptions } from './types';
 import { SheetOf } from '../types';
 
@@ -317,52 +314,6 @@ export class PlustekScanner implements Scanner {
       delete this.statusOverride;
     }
   }
-}
-
-const PutMockRequestSchema = z.object({
-  files: z.tuple([z.string(), z.string()]),
-});
-
-export function plustekMockServer(client: MockScannerClient): Application {
-  return express()
-    .use(express.raw())
-    .use(express.json({ limit: '5mb', type: 'application/json' }))
-    .use(express.urlencoded({ extended: false }))
-    .put('/mock', async (request, response) => {
-      const bodyParseResult = safeParse(PutMockRequestSchema, request.body);
-
-      if (bodyParseResult.isErr()) {
-        response
-          .status(400)
-          .json({ status: 'error', error: `${bodyParseResult.err()}` });
-        return;
-      }
-
-      const simulateResult = await client.simulateLoadSheet(
-        bodyParseResult.ok().files
-      );
-
-      if (simulateResult.isErr()) {
-        response
-          .status(400)
-          .json({ status: 'error', error: `${simulateResult.err()}` });
-        return;
-      }
-
-      response.json({ status: 'ok' });
-    })
-    .delete('/mock', async (_request, response) => {
-      const simulateResult = await client.simulateRemoveSheet();
-
-      if (simulateResult.isErr()) {
-        response
-          .status(400)
-          .json({ status: 'error', error: `${simulateResult.err()}` });
-        return;
-      }
-
-      response.json({ status: 'ok' });
-    });
 }
 
 export function withReconnect(
